LS: fix panic in lsp_helper and move it to LsSemanticGroup
