+++ crates/cairo-lang-language-server/src/ide/completion/completions.rs
@@ -24,6 +24,7 @@ use tower_lsp::lsp_types::{CompletionItem, CompletionItemKind, Position, Range, 
 use tracing::debug;
 
 use crate::ide::utils::find_methods_for_type;
+use crate::lang::db::semantic_extension::LsSemanticGroupExtension;
 use crate::lang::db::{AnalysisDatabase, LsSemanticGroup};
 use crate::lang::lsp::ToLsp;
 
@@ -274,7 +275,7 @@ pub fn completion_for_method(
 
     // If the trait is not in scope, add a use statement.
     if !module_has_trait(db, module_id, trait_id)? {
-        if let Some(trait_path) = db.visible_traits_from_module(module_id).get(&trait_id) {
+        if let Some(trait_path) = db.visible_traits_from_module(module_id).ok()?.get(&trait_id) {
             additional_text_edits.push(TextEdit {
                 range: Range::new(position, position),
                 new_text: format!("use {};\n", trait_path),
