+++ crates/cairo-lang-semantic/src/lib.rs
@@ -2,6 +2,11 @@
 //! The semantic model represents the Cairo program after type resolution and some syntax
 //! desugaring.
 
+pub use diagnostic::SemanticDiagnostic;
+pub use substitution::SemanticObject;
+
+pub use self::semantic::*;
+
 pub mod corelib;
 pub mod db;
 pub mod diagnostic;
@@ -10,7 +15,6 @@ pub mod inline_macros;
 pub mod items;
 pub mod literals;
 pub mod lookup_item;
-pub mod lsp_helpers;
 pub mod plugin;
 pub mod resolve;
 pub mod substitution;
@@ -18,11 +22,6 @@ pub mod types;
 pub mod usage;
 
 mod semantic;
-
-pub use diagnostic::SemanticDiagnostic;
-pub use substitution::SemanticObject;
-
-pub use self::semantic::*;
 
 #[cfg(any(feature = "testing", test))]
 pub mod test_utils;
